requirements.txt
asgiref==3.2.10
Django==3.1.2
django-phonenumber-field==5.0.0
djangorestframework==3.12.1
phonenumbers==8.12.10
pytz==2020.1
sqlparse==0.3.1
